+++ src/main/java/com/kakaotech/team14backend/point/infrastructure/PointHistoryRepository.java
@@ -1,10 +1,11 @@
-package com.kakaotech.team14backend.inner.point.repository;
+package com.kakaotech.team14backend.point.infrastructure;
 
-import com.kakaotech.team14backend.inner.point.model.PointHistory;
-import java.util.List;
-import java.util.Optional;
+import com.kakaotech.team14backend.point.domain.PointHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+
+import java.util.List;
+import java.util.Optional;
 
 public interface PointHistoryRepository extends JpaRepository<PointHistory, Long> {
